exception handling and restDocs(for clients,memberships) implemented

+++ src/main/java/com/exadel/fedorov/orders/controllers/ClientController.java
@@ -3,6 +3,8 @@ package com.exadel.fedorov.orders.controllers;
 import com.exadel.fedorov.orders.domain.Client;
 import com.exadel.fedorov.orders.dto.dto_request.NewClientDTO;
 import com.exadel.fedorov.orders.dto.dto_response.ClientDTO;
+import com.exadel.fedorov.orders.exception.InvalidDataException;
+import com.exadel.fedorov.orders.exception.NoSuchDataException;
 import com.exadel.fedorov.orders.service.ClientService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -24,28 +26,32 @@ import java.util.Optional;
 @RestController
 public class ClientController {
 
-    public static final String SUCH_CLIENT_DOES_NOT_EXIST_MESSAGE = "";//TODO
-    public static final String NO_CLIENTS_FOUND_MESSAGE = "";
+    private static final String SUCH_CLIENT_DOES_NOT_EXIST_MESSAGE = "Such client does not exist.";
+    private static final String NO_CLIENTS_FOUND_MESSAGE = "No clients found.";
+    private static final String CLIENT_ID_IS_NOT_VALID = "Client Id is not valid";
 
     @Autowired
     private ClientService clientService;
 
     @GetMapping
-    public ResponseEntity<List<ClientDTO>> readAll() {
+    public ResponseEntity<List<ClientDTO>> readAll() throws NoSuchDataException {
         List<ClientDTO> clientDTOS = clientService.findAll();
         if (clientDTOS.isEmpty()) {
-            return new ResponseEntity(NO_CLIENTS_FOUND_MESSAGE, HttpStatus.NOT_FOUND);
+            throw new NoSuchDataException(NO_CLIENTS_FOUND_MESSAGE);
         }
-        return new ResponseEntity(clientDTOS, HttpStatus.OK);
+        return new ResponseEntity<>(clientDTOS, HttpStatus.OK);
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<ClientDTO> read(@PathVariable("id") Long id) {
-        if (id == null) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    public ResponseEntity<ClientDTO> read(@PathVariable("id") Long id) throws NoSuchDataException, InvalidDataException {
+        if (null == id || id <= 0) {
+            throw new InvalidDataException(CLIENT_ID_IS_NOT_VALID);
         }
         Optional<ClientDTO> clientDTO = clientService.findById(id);
-        return clientDTO.map(respOrderDTO -> new ResponseEntity(clientDTO, HttpStatus.OK)).orElseGet(() -> new ResponseEntity(SUCH_CLIENT_DOES_NOT_EXIST_MESSAGE, HttpStatus.NOT_FOUND));
+        if (!clientDTO.isPresent()) {
+            throw new NoSuchDataException(SUCH_CLIENT_DOES_NOT_EXIST_MESSAGE);
+        }
+        return new ResponseEntity<>(clientDTO.get(), HttpStatus.OK);
     }
 
     @ResponseStatus(HttpStatus.CREATED)
@@ -57,14 +63,20 @@ public class ClientController {
 
     @ResponseStatus(HttpStatus.OK)
     @PutMapping(value = "/{id}")
-    public ResponseEntity<Client> update(@PathVariable("id") Long id, @RequestBody NewClientDTO updateClient) {
+    public ResponseEntity<Client> update(@PathVariable("id") Long id, @RequestBody NewClientDTO updateClient) throws InvalidDataException {
+        if (null == id || id <= 0) {
+            throw new InvalidDataException(CLIENT_ID_IS_NOT_VALID);
+        }
         clientService.update(id, updateClient);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping(value = "/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<Client> deleteClient(@PathVariable("id") Long id) {
+    public ResponseEntity<Client> deleteClient(@PathVariable("id") Long id) throws InvalidDataException {
+        if (null == id || id <= 0) {
+            throw new InvalidDataException(CLIENT_ID_IS_NOT_VALID);
+        }
         clientService.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
