+++ src/main/java/com/exadel/fedorov/orders/controllers/OrderController.java
@@ -3,6 +3,8 @@ package com.exadel.fedorov.orders.controllers;
 import com.exadel.fedorov.orders.domain.Order;
 import com.exadel.fedorov.orders.dto.dto_request.ReqOrderItemDTO;
 import com.exadel.fedorov.orders.dto.dto_response.RespOrderDTO;
+import com.exadel.fedorov.orders.exception.InvalidDataException;
+import com.exadel.fedorov.orders.exception.NoSuchDataException;
 import com.exadel.fedorov.orders.service.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -29,33 +31,37 @@ public class OrderController {
     private static final String SUCH_ORDER_DOES_NOT_EXIST_MESSAGE = "Such order does not exist.";
     private static final String NO_ORDERS_FOUND_MESSAGE = "No orders found.";
     private static final String INCORRECT_INPUT_DATA_MESSAGE = "Incorrect input data.";
+    private static final String ORDER_ID_IS_NOT_VALID = "Order Id is not valid";
 
     @Autowired
     private OrderService orderService;
 
     @GetMapping
-    public ResponseEntity<List<RespOrderDTO>> readAll() {
+    public ResponseEntity<List<RespOrderDTO>> readAll() throws NoSuchDataException {
         List<RespOrderDTO> orders = orderService.findAll();
         if (orders.isEmpty()) {
-            return new ResponseEntity(NO_ORDERS_FOUND_MESSAGE, HttpStatus.NOT_FOUND);
+            throw new NoSuchDataException(NO_ORDERS_FOUND_MESSAGE);
         }
-        return new ResponseEntity(orders, HttpStatus.OK);
+        return new ResponseEntity<>(orders, HttpStatus.OK);
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<RespOrderDTO> read(@PathVariable("id") Long orderId) {
-        if (orderId == null) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    public ResponseEntity<RespOrderDTO> read(@PathVariable("id") Long orderId) throws NoSuchDataException, InvalidDataException {
+        if (null == orderId || orderId <= 0) {
+            throw new InvalidDataException(ORDER_ID_IS_NOT_VALID);
         }
         Optional<RespOrderDTO> orderDto = orderService.findById(orderId);
-        return orderDto.map(respOrderDTO -> new ResponseEntity(respOrderDTO, HttpStatus.OK)).orElseGet(() -> new ResponseEntity(SUCH_ORDER_DOES_NOT_EXIST_MESSAGE, HttpStatus.NOT_FOUND));
+        if (!orderDto.isPresent()) {
+            throw new NoSuchDataException(SUCH_ORDER_DOES_NOT_EXIST_MESSAGE);
+        }
+        return new ResponseEntity<>(orderDto.get(), HttpStatus.OK);
     }
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
-    public ResponseEntity<Order> create(@RequestBody List<ReqOrderItemDTO> orderItems) {
+    public ResponseEntity<Order> create(@RequestBody List<ReqOrderItemDTO> orderItems) throws InvalidDataException {
         if (orderItems.isEmpty()) {
-            return new ResponseEntity(INCORRECT_INPUT_DATA_MESSAGE, HttpStatus.BAD_REQUEST);
+            throw new InvalidDataException(INCORRECT_INPUT_DATA_MESSAGE);
         }
         orderService.createOrder(orderItems, DEFAULT_CLIENT_NAME, DEFAULT_STATUS_DESCRIPTION);
         return new ResponseEntity<>(HttpStatus.CREATED);
@@ -63,14 +69,22 @@ public class OrderController {
 
     @ResponseStatus(HttpStatus.OK)
     @PutMapping(value = "/{id}")
-    public ResponseEntity<Order> update(@PathVariable("id") Long id, @RequestBody String status, @RequestBody String statusDescription) {
+    public ResponseEntity<Order> update(@PathVariable("id") Long id,
+                                        @RequestBody String status,
+                                        @RequestBody String statusDescription) throws InvalidDataException {
+        if (null == id || id <= 0) {
+            throw new InvalidDataException(ORDER_ID_IS_NOT_VALID);
+        }
         orderService.update(id, status, statusDescription);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping(value = "/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<Order> deleteOrder(@PathVariable("id") Long id) {
+    public ResponseEntity<Order> deleteOrder(@PathVariable("id") Long id) throws InvalidDataException {
+        if (null == id || id <= 0) {
+            throw new InvalidDataException(ORDER_ID_IS_NOT_VALID);
+        }
         orderService.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
